refactor(ai): share native OpenAI client construction
